Dependency installer intended for homeworks at soyhenry.com
This script will angrily look for folders named "homework" that have a "package.json" file and perform a simple "npm install" operation on them. It's also commented on painful detail, just in case some misguided soul stumbles upon it and try to use it to learn bash.
Written with love and closely following the spirit of Larry Wall's Three Cardinal Virtues of a programmer: Laziness, Impatience, and Hubris. (https://avdi.codes/the-three-virtues/)